+++ src/backend/main/go.backends/LDAInterfaces.go
@@ -4,7 +4,11 @@
 
 package backends
 
+import "github.com/CaliOpen/CaliOpen/src/backend/defs/go-objects"
+
 type LDABackend interface {
 	GetRecipients([]string) ([]string, error)
 	StoreRaw(data string) (raw_id string, err error)
+	GetMessage(user_id, msg_id string) (msg *objects.MessageModel, err error)
+	UpdateMessage(msg *objects.MessageModel, fields map[string]interface{}) error
 }