+++ src/frontend/web_application/src/components/ContactAvatarLetter/index.jsx
@@ -8,7 +8,11 @@ export const SIZE_MEDIUM = 'medium';
 export const SIZE_LARGE = 'large';
 export const SIZE_XLARGE = 'xlarge';
 
-const ContactAvatarLetter = ({ contact, size }) => {
+const ContactAvatarLetter = ({ contact, size, isRound }) => {
+  const ClassNameShape = classnames(
+      'm-avatar--round': isRound,
+  );
+
   const classNameModifiers = {
     [SIZE_SMALL]: 'm-avatar--small',
     [SIZE_MEDIUM]: 'm-avatar--medium',
@@ -26,7 +30,7 @@ const ContactAvatarLetter = ({ contact, size }) => {
   const letterClassNameSize = classNameSize ? `${classNameSize}__letter` : null;
 
   return (
-    <div className={classnames('m-avatar', classNameSize)}>
+    <div className={classnames('m-avatar', classNameSize, ClassNameShape)}>
       <ContactIconLetter
         className={classnames('m-avatar__letter', letterClassNameSize)}
         contact={contact}
@@ -38,6 +42,7 @@ const ContactAvatarLetter = ({ contact, size }) => {
 ContactAvatarLetter.propTypes = {
   contact: PropTypes.shape({}),
   size: PropTypes.string,
+  isRound: PropTypes.bool,
 };
 
 export default ContactAvatarLetter;