+++ src/frontend/web_application/src/components/form/index.js
@@ -1,4 +1,3 @@
-// import CheckboxField from './';
 import FieldErrors from './FieldErrors';
 import Fieldset, { Legend } from './Fieldset';
 import FormGrid, { FormRow, FormColumn } from './FormGrid';
@@ -7,11 +6,11 @@ import PasswordStrength from './PasswordStrength';
 import RadioFieldGroup from './RadioFieldGroup';
 import SelectFieldGroup from './SelectFieldGroup';
 import CheckboxFieldGroup from './CheckboxFieldGroup';
+import CollectionFieldGroup from './CollectionFieldGroup';
 import TextFieldGroup from './TextFieldGroup';
 import TextareaFieldGroup from './TextareaFieldGroup';
 
 export {
-  // CheckboxField,
   FieldErrors,
   Fieldset,
   Legend,
@@ -23,6 +22,7 @@ export {
   PasswordStrength,
   SelectFieldGroup,
   CheckboxFieldGroup,
+  CollectionFieldGroup,
   TextFieldGroup,
   TextareaFieldGroup,
 };