+++ src/frontend/web_application/stories/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [2017-02-27]
+
 ### Added
 
   - Settings - DevicesManagement
@@ -9,11 +11,16 @@
   - Pi - PiBar
   - Layout - Section
   - Lists - DefList
+  - form - CollectionFieldGroup
 
 ### Changed
 
   - Form - Field with error: no more uppercase for the message & replace background color by a red border (inset)
   - Button - Add "inline" prop, to be used for Button placed beside inputs (dark background, 1px left margin, small font)
+
+### Removed
+
+  - BlockList - background color and ItemContent have been removed
 
 ## [2017-01-19]